feat: updated docs, updated validated commit tests
