+++ docs/updater/update_process.md
@@ -57,6 +57,7 @@ specified (similarly to how `git pull` works)
 - `--format-output`: Return formatted output including information on whether the build was successful and an error message if it was raised.
 - `--exclude-target`: Globs defining which target repositories should be ignored during the update. Accepts multiple values.
 - `--strict`: Enable/disable strict mode. In strict mode, an error is returned if warnings are raised. Default is disabled.
+- `--force`: Flag used to run a forced update.
 
 ### Determining filesystem paths of repositories
 
@@ -158,3 +159,12 @@ if __name__ == '__main__':
     state = do_something(data)
     send_state(state)
 ```
+
+
+# Forced Updates
+
+A forced update overrides the normal update process to address situations where the authentication repository or its target repositories have encountered inconsistencies that prevent a standard update. During a forced update, the updater bypasses certain checks that would normally stop the process, such as the presence of an invalid or missing last validated commit. The forced update ensures that the repository is aligned with the latest commit found on the remote repository. 
+
+By using the --force flag, the updater will reset the state of the repository, discarding any uncommitted changes, and align the repository with the latest commit from the remote repository. This includes resetting the last validated commit to the latest commit found on the remote, and synchronizing the target repositories accordingly. If the target repositories have diverged or contain unaligned commits, the updater will reset them to the validated state, potentially discarding local changes.
+
+However, if the remote repository itself is in an inconsistent state—such as having missing critical metadata, or if the commit history has been rewritten, the forced update may still fail. For example, if a file is updated but not committed in the authentication repository, this creates a "dirty index" which makes the updater fail.
